chapter6/salesitemservice/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
ariadne==0.22
black==22.1.0
blue==0.9.1
click==8.1.7
dependency-injector==4.41.0
dnspython==2.6.1
fastapi==0.110.0
flake8==4.0.1
googleapis-common-protos==1.62.0
graphql-core==3.2.3
grpcio==1.62.1
grpcio-status==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
httptools==0.6.1
idna==3.6
libcst==1.2.0
markdown-it-py==3.0.0
mccabe==0.6.1
mdurl==0.1.2
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
pathspec==0.12.1
platformdirs==4.2.0
protobuf==4.25.3
pycodestyle==2.8.0
pydantic==2.6.3
pydantic_core==2.16.3
pyflakes==2.4.0
Pygments==2.17.2
pymongo==4.6.2
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
strawberry-graphql==0.220.0
tomli==2.0.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
